requirements.txt
# 🎥 SHUJAA STUDIO - PRODUCTION REQUIREMENTS
# 🏆 100% InVideo Parity + African AI Video Generation
# ⚡ Optimized for fast startup and production deployment

# Core AI Video Generation Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
gradio>=4.0.0

# Audio Processing & Speech (Production Ready)
av>=13.0.0
faster-whisper>=0.9.0
# openai-whisper>=20231117
git+https://github.com/suno-ai/bark.git
pyttsx3>=2.90
soundfile>=0.12.1

# Image Processing & Generation (SDXL Ready)
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.16.0

# Watermark Removal Dependencies

# Model Management & Performance (Lazy Loading Support)
accelerate>=0.20.0
safetensors>=0.3.0
huggingface-hub>=0.14.1
tokenizers>=0.21.0
xformers>=0.0.22; platform_system != "Windows"  # Memory optimization

# AI Model Dependencies  
einops>=0.8.0
funcy>=2.0
encodec>=0.1.1

# Voice & Music for InVideo Competition  
edge-tts>=7.2.0
scipy>=1.14.0

# Advanced Media Processing (Phase 5 Complete)
librosa>=0.10.0  # Music analysis & mixing
opencv-python>=4.8.0  # Video processing
opencv-contrib-python>=4.8.0  # Extended CV features

# Professional Video Features (Production Ready)
# face-recognition>=1.3.0  # Avatar support
# dlib>=19.24.0  # Face detection
requests>=2.28.0
beautifulsoup4>=4.12.2
gtts>=2.3.2
feedparser>=6.0.10
huggingface_hub>=0.17.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pyyaml>=6.0.1
dotmap>=1.3.30
redis>=4.5.0
boto3>=1.28.0
celery>=5.3.4
aiohttp>=3.8.5
fastapi>=0.103.0 # For API framework
uvicorn[standard]>=0.23.2
pyjwt>=2.8.0 # For JWT handling
cryptography>=41.0.3
SQLAlchemy>=2.0.0 # For database ORM
passlib[bcrypt]>=1.7.4
fastapi-limiter>=0.1.5 # For API rate limiting
python-dotenv>=1.0.0  # Environment management


# Multi-Language Support
Babel>=2.12.1

# Performance & Deployment (Optimization Phase)
psutil>=5.9.0  # System monitoring  
tqdm>=4.65.0  # Progress bars
colorama>=0.4.6  # Terminal colors
click>=8.1.0  # CLI interface

py-spy>=0.3.14

# Optional: Advanced AI Features (Install when needed)
# git+https://github.com/Rudrabha/Wav2Lip.git  # Lip-sync
# git+https://github.com/OpenTalker/SadTalker.git  # Talking heads
# git+https://github.com/facebookresearch/audiocraft.git  # Music generation
# controlnet-aux>=0.4.0  # ControlNet support for precise image control
